requirements.txt
# Fortress Options — Python Requirements
yfinance>=0.2.37
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.1
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
stripe>=8.0.0
python-multipart>=0.0.9
firebase-admin>=6.4.0
anthropic>=0.40.0
psycopg2-binary>=2.9.9
# Used as the yfinance HTTP session to dodge rate limits via Chrome impersonation.
curl_cffi>=0.7.0
